requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==0.63.3
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.6
oauthlib==3.2.2
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
